Fix nitro implementation
The allocations table has been renamed to holdings.

Guarantees and allocations are both type Outcome now.
(The non-zeroness of the guaranteedChannel indicates the type of
outcome.)
This means that the ForceMove API will "just work" with guarantees.

Reprioritize was corrected to allow different lengths in
allocation.destination vs guarantee.destination.
(allocation.destination must be at least as long as
guarantee.destination.)