WIP: Remove dependency on ethers.utils.BigNumber
ethers.utils.BigNumber causes a lot of headaches.

For example, see the discussion here: https://gitter.im/ethers-io/Lobby?at=5c69a8394bafd97ca30a8940

In addition, specifying BigNumber as the type of a number on an
interface goes against the ethos that front-end apps should pass around
objects that can be stored in a redux store, ie. vanilla JSON objects.

An ethers contract will accept a number as an argument for a smart
contract function that expects a uint256 solidity variable. However, it
returns a `BigNumber`.

Therefore, we change the type on the smart contract to uint32 for all
but the allocation attribute. ethers will return these numbers as
a number in javascript.

The allocation attribute is a Uint256[] array, which is now an alias for
a string[]. ie, we expect the allocation to be stored in its hexstring
format.

WIP because ethers' implementation of BigNumber.toHexString does this:
https://github.com/ethers-io/ethers.js/blob/51fb4728091a30d147b7dcd61e08562c6171b372/src.ts/utils/bignumber.ts#L30

which I don't yet understand.