Refactor ConsensusApp
This commit removes references to the updateType, and therefore does not
select the action based on the updateType.

Instead, `validateActionX` returns false if the transition is not
`ActionX`. If the transition is `ActionX`, then checks are performed,
and `true` is returned.

This allows `validTransition` to return
```
validatePropose(oldCommitment, newCommitment, numParticipants)   ||
validateVote(oldCommitment, newCommitment) ||
validateVeto(oldCommitment, newCommitment) ||
validatePass(oldCommitment, newCommitment) ||
validateFinalVote(oldCommitment, newCommitment) ||
invalidTransition();
```