Allow proposedAllocation to be empty
This allows guarantor channels to use the consensus app.

This offloads some responsibility onto the client to ensure they propose
a valid outcome. However, the core rules still prevent the final vote
transition from being valid in case someone proposes an empty allocation
in a ledger channel.

The validTransition function is not updated, since we plan to delete it.