Use Test{Eth,Erc}20AssetHolders in NitroProtocol tests
Currently the NetworkContext exports a dict of contract names to addresses.
This leads to a problem if we want to share the same set of deployments
across packages, but also have the EthAssetHolder hold a reference to the
TestNitroAdjudicator in the nitro tests and hold a reference to the
NitroAdjudicator in the wallet/hub packages (because the same name would
point to multiple deployed contracts). This commit avoids that problem by
by prefixing the names of the AssetHolders that point to the TestNitroAdjudicator
with Test*.