ChannelStoreEntry does not expose references to store data
Previously, `allocateToTarget` from `calculations.ts` would have
a reference to a value stored in the store. The implementation of
`allocateToTarget` was modifying the contents of an array in place,
which had the effect of modifying a state after the fact.

No code should get direct access to someStore._store[channelId].
Instead, it should call someStore.getEntry, which only gets references
to deep copies of that data.