Add progress bar
This is a bit hacky but tricky to do better when we can't access more information about the invoked machine(s). Spawning machines as actors may help, as we can then sync state with the parent https://xstate.js.org/docs/guides/actors.html#syncing-and-reading-state . The progress information could be stored in the metadata of each state in the machine config.