Remove budgets from workflows
At this point, we do not know how budgets should work. For example,
consider this scenario:
Start:
```
{
    free: { hub: 10, player: 15 },
    inUse: { hub: 0, player: 0 }
}
```
Fund a virtual channel with `{ hub: 4, player: 6}`
```
{
   free: { hub: 6, player: 9 },
   inUse: { hub: 4, player: 6 },
}
```
Virtual channel terminates, ends in `{hub: 9, player: 1}`
```
{
   free: { hub: 10, player: 10 },
   inUse: { hub: 0, player: 0 },
}
```
Where does the extra 5 go?
Do we store it in an `excess` field?

Or we add it to the `free.hub` field, even though the user has not approved
receiving more than 10 funds.

It seems like we need one or the other -- otherwise the `free.hub
+ inUse.hub` amount will decrease over time.

While we decide this, putting the extra 5 in the `hub.free` field seems to make
the most sense. At that point, the budget is simply a projection of the
current outcome of the ledger channel with the hub.

Keping both things in sync in the store, and using locks to do this
safely with concurrent channels, seems like too much complexity right
now given that the budgets have no benefit. Therefore, this commit
removes budget-related details from the workflows.