Switch from concat to merge in chainUpdatedFeed
I think `concat` has a nasty concurrency edge case that wouldn't rear
its head in practice, but would in tests:
A. if you `concat(first, updates)`, then I think that the `first` promise
has to resolve before the `updates` start listening for events.
B. If you `concat(updates, first)` then you would not get the `first`,
since `updates` has to finalize first

In practice, A is a pretty minor issue because of real timings, but it
was causing a test failure due to the indexDB changes.