Remove a bunch more uses of npx
Only need to use npx if the command needs to be hunted around for before executing: i.e. when the command does not exist in this package (it might be in another package, or on npm).

Downsides of npx: there is an overhead involved in searching for the command, and it could obscure incorrect dependency specifications.