Consolidate UpdateChannel and SignState actions
... into one (more explicit) action

It is important for `UpdateChannel` to specify all state variables --
otherwise, who decides which state variables to use? Should it come
from the latest state? the supported state? My latest state?

It seems like it's the protocol's responsibility to answer that.

When all state variables must be given, UpdateChannel and SignState
both mean the same thing: "Sign this state".