Improve lockApp error handling
It seems like the default behaviour when a channel is missing should be
to throw a generic ChannelMissing error. This attempts to minimize the
chance that a missing channel goes unhandled.

This allows you to pass a custom missing channel handler, in case you
want to throw a different error.