feat: introduce Address and PrivateKey nominal types
Some fields stored by the wallet and transferred via the wire protocol are
bytes that encode hexadecimal values. For those fields, we will try to
lower case the values (as opposed to checksum the values)

The primary goal of address checksums is to prevent user error when inputting
ethereum addresses. As the wallet never modifies an address, checksumming
will not prevent wallet errors.

Consequently, it is more consistent to apply the heuristic that all hexadecimal
values should be lower case.