feat: add fundingStatus to ChannelResult
Note that, for a single channel state, the fundingStatus might be
different for different participants. For example, consider a channel
with an outcome of 5 eth allocated to each Alice and Bob. Consider that
the channel has on-chain funding of 0 eth. For Alice, the funding status
is ReadyToFund. For Bob, fundingStatus is Uncategorized.