fix: clean up a race conditions with finalizing channels
Before this commit, the strategy was to use a pqueue to limit the number
of onBlockMined callbacks executing simultaneously to one. This doesn't
quite work as there are other places in the chain service that mutate the
finalizingChannels list.

The new approach is to always remove an element from finalizingChannels
before operating on the element.