test: do not use arbitrary, constant block timestamps in tests
At the moment, every test case only mines the blocks that are needed for
the testcase. As a result, we should NOT need to mine a block with a timestamp
very far in the future in order to have a finalizing channel on chain
with finalizesAt greater than the current block timestamp.

Moreover, mining a block with a constant timestamps might lead to subtle test issues.
Consider two tests cases A and B. A mines a blocks with timestamp T1 and T2
to create a test environment necessary to assert a behavior. B mines T3 and
T4.

We do not know whether A or B will run first. As a result, it might not
be possible to determine T1, T2, T3, T4 where A and B will pass no matter
which test is run first.

An example of the above is:
- A sets channel storage with finalizesAt of 2.
- A mines a block with timestamp 1.
- A asserts that channel is not finalized.
- A mines a block with timestamp 2.
- A asserts that the channel is finalized.

- B sets channel storage with finalizesAt of 4.
- B mines a block with timestamp 3.
- B asserts that channel is not finalized.
- B mines a block with timestamp 4.
- B asserts that the channel is finalized.

A test case will no longer be valid if executed after B.