refactor: make fundingStatus optional in the ChannelResult
Currently the fundingStatus value is sometimes inaccurate. For example,
it's possible to get 'Uncategorized' or 'ReadyToFund' when the
channel is funded, but the funding hasn't been fetched from the db.

Fixing this would require always querying the db for funding, which
might not be desireable for performance reasons. Instead, consumers
will now receive fundingStatus: undefined if it hasn't be fetched.