refactor: deduplicate validTransition check
Note that the existing validChain function returned false when the appData was '0x00'. There are some issues with this:
- string equality to '0x00' is not a robust check for zero
- null apps have a zero appDefinition, not a zero appData

It was also necessary to introduce isNullApp getter. Some of our tests (namely, process-ledger-queue.test.ts) weren't setting the asset holder on outcomes for ledger channels. The existing isLedger getter therefore returned false and confusion followed.