chore: add a pull_request template to make our code reviewers love us
Allows developers to use this file as a template when creating a PR.
Inspired by
- https://mtlynch.io/code-review-love/
- https://google.github.io/eng-practices/review/developer/cl-descriptions.html

I think that the mandatory sections of this template provide enough value
to both the PR reviewer and to future developers, to merit their inclusion
in the vast majority of PRs.

The optional sections serve as good reminders, when appropriate.

Using a strict template gives us some additional benefits. For example,
we could write a github action that adds a "Breaking change" label whenever
the PR description contains the string `'- [x] Breaking change'`.