chore: fail or succeed a challenge objective
Before this change, there is a set of "warn" conditions for the challenge
objective. With these conditions, the objective will neither fail nor succeed,
likely resulting in every API call to the wallet inspecting the objective
and concluding that it is in the warn state.

This change forces the objective to either fail or succeed.