refactor: do not rely on directFundingStatus internally
directFundingStatus of ChannelState has likely outlived its usefullness.
The original intention with directFundingStatus was to create a function
that maps a channel + funding state to a channel funding string.

We are seeing the channel opener and defunding protocols reimplement parts
of the directFundingStatus. That's probably the right approach. This way
we can build up smaller functions that can answer specific questions such
as:
- is this directly funded channel fully funded?
- do I have any funds in this channel that I can claim?