test: remove fundStatus check in directly-funded-channel.test.ts
fundingStatus is an unrealiable part of the channel result. fundingStatus is
undefined if the channels table is not joined with the funding table
during protocol execution. The channels table is joined with the funding
table only when a protocol needs funding data.

It seems that fundingStatus has outlived its usefullness and should be
removed (especially now that we are moving toward first-class objectives)