fix: do not assume that a ledger channel is directly funded
At the moment, a ledger channel might be directly funded or fake funded.
In the future, a ledger channel might be virtually funded or funded by
another ledger channel.

In any of the above cases, when funding a channel with a ledger channel,
we should only proceed with the ledger funding protocol if the ledger
channel is in the running stage regardless of how it was funded.