refactor: create a transaction submission order instead of a primary submitter
If a participant has no funds in a channel and the participant is the
primary transaction submitter, it is not clear which participant should
submit collaborative transactions. The primary participant has no incentive
to pay the gas fees.

With a prioritized list of transaction submitters the logic becomes:
1. Participants agree on the order of transaction submitters when creating
the close channel objective.
2. When the channel is defunded, the highest priority submitter who has
funds in the channel submits the collaborative transactions.

An alternative approach is to determine the participant who will submit
collaborative transactions when creating the close channel objective by
inspecting current channel allocations. But it might be possible for the channel
allocations to change between close channel objective creation and defunding.