chore: fix blockConfirmations inconsistency
At the moment, the chain service defines number of confirmations as
the number of blocks mined AFTER the transaction block. A likely more
intuitive definition is for number of confirmations to be defined as
the number of blocks that INCLUDE the transaction block.

The bug below was a result of mixing the definitions.