test: remove mineBlocksSlow
The following properties are nice to have for the chain service testsuite:
1. A test does not mine blocks after completion.
2. A test runs as quickly as possible.

Given the above, each test should only mine the minimum number of blocks
needed for a test to pass.

In addition, when possible, it is nice for a test to be written so that the test
execute as deterministically as possible. This leads to the following pattern:
1. Set up chain service and maybe a certain contract state.
2. Mine a few blocks so that some contract state remains unconfirmed.
     Check that certain constraints hold.
2. Mine a few more blocks so that contract state is confirmed.
     Check that a certain new set of constraints holds.