chore: add an an address parameter to ChainWatcher constructor
The chain watcher relies on a selected address to sign transactions:
     return this.provider.getSigner(this.selectedAddress as string);

In the browser environment, the selected address is read from window.ethereum.
The address constructor parameter in the chain watcher lets us use the chain
watcher without a window.ethereum object defined.

This has been helpful in writing interop tests where the channel wallet
is instantiated outside of the browser.