test: correctly construct the ledger variables after sync cranking
I am aiming towards simpler setup & assertions that does not require
database access. I need to ensure that I am not changing the
behaviour of the test. So, I am asserting that the thing that I want to
test, is equal to the thing that the current test asserts is "correct".