refactor, perf(test): test synchronous LedgerManager logic (#3440)
Currently, ledger-manager.test.ts tests LedgerManager.crank, which is asynchronous and requires DB access.

This PR exposes the synchronous logic (which accounts for ~90% of the logic of LedgerManager) as a public method, and incrementally refactors testLedgerCrank to be a synchronous function. The synchronous logic is purish -- it does not mutate the Channel model, but does mutate the LedgerRequests[] models that are passed in.

The main motivation of this work is to move the ledger funding protocol into wallet-core, where it can be re-used by the xstate-wallet package. Secondary benefits are:
- the tests are significantly faster: with no DB access, I observed a 50% speedup locally. Since it previously took ~60s to run the entire test suite, this is a significant boon to anyone updating the ledger funding logic.
- the tests do not hit the DB: we can run parallel tests again!
- the tests are easier to debug: I do not have good success debugging tests that have asynchronous DB access. Pausing the process tends to cause knex timeouts.

The shortcomings of this approach are that the database access used by LedgerManager.crank are not tested in ledger-manager.test.ts. Since this logic is tested in our with-peers integration test, I believe this is an acceptable tradeoff.


### How Has This Been Tested?

My approach here was to figure out how to accurately construct the test input and output without introducing "regressions". So, I pass through many intermediate states where the test is testing itself against the old behaviour.

I believe the ledger-manager.test.ts passes on every commit.


### Supporting Changes

* refactor: TestLedgerChannel returns LedgerRequest models synchronously, without inserting them into the store

* getAndLockChannel always returns a Channel. 

* refactor: getAndLockChannel always returns a channel
  
currently, every callee immediately checks that the return value is truthy, and throws elsewise.
this saves every caller from having to perform the same check

* refactor: LedgerStateDesc -> LedgerStateDescription

I was confused, thinking LedgerStateDesc had something to do with descending

* refactor: test-ledger-channel methods return a RichLedgerRequest

* refactor: :tada: don't use LedgerRequest in ledger-manager test

* test: ledger-manager test is sync