refactor: collapse approve objective into crank objective function
As of https://github.com/statechannels/statechannels/pull/3505, crankRichObjectives
is a generic handler for any objective event. approveRichObjectives was
written with custom logic for the approve event (the custom logic sends
the objective to other participants). This approve event logic will not
get triggered for events arriving directly to crankRichObjectives.

This commit moves the custom logic into the crankRichObjectives function.