refactor: introduce and use RichObjective and RichObjectiveEvent types
Browser wallet is the only consumer of rich objectives at the moment.
Browser wallet hardcoded the following assumptions:
- Only one rich objective in the store.
- Only consider the direct funder open channel objective.

This commit introduces the RichObjective and RichObjectiveEvent abstractions
to help the browser wallet remove the above assumptions. At the moment,
a RichObjective is always DirectFunder.OpenChannelObjective and a RichObjectiveEvent
is always a DirectFunder.OpenChannelEvent