properly crank objectives in chain event listeners The main focus of this commit is to get the chain events to properly trigger the crank and deal with the result.  They now all follow the basic pattern of updating the store, cranking the engine and dealing with the result.
Another change is that close channels now creates its own promise for
done. I need to dig into this further but it likely indicates that the
ensureObjectives function is not working like it should.

The create chain listener function has been moved  to be a function of the
wallet so it can easily access wallet internals.